requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
Django==5.1.1
django-autoslug==1.9.9
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-ninja==1.3.0
django-querycount==0.8.3
dnspython==2.6.1
email_validator==2.2.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
psycopg==3.2.2
psycopg-binary==3.2.2
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-decouple==3.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
whitenoise==6.7.0
